add midlleware

+++ routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\CompanyController;
+use App\Http\Middleware\isCompanyOwner;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
@@ -16,8 +17,8 @@ Route::controller(AuthController::class)->group(function() {
 });
 
 Route::controller(CompanyController::class)->group(function() {
-    Route::post("companies/{company_id}/invites", "createInviteCode");
-    Route::get("companies/{company_id}/invites", "getAllInviteCodes");
-    Route::delete("companies/{company_id}/invites/{invite_id}", "deleteCompanyInviteCode");
-    Route::get("companies/{company_id}/users", "getAllUsers");
+    Route::post("companies/{company_id}/invites", "createInviteCode")->middleware(['auth:sanctum',isCompanyOwner::class]);
+    Route::get("companies/{company_id}/invites", "getAllInviteCodes")->middleware(['auth:sanctum',isCompanyOwner::class]);
+    Route::delete("companies/{company_id}/invites/{invite_id}", "deleteCompanyInviteCode")->middleware(['auth:sanctum',isCompanyOwner::class]);
+    Route::get("companies/{company_id}/users", "getAllUsers")->middleware(['auth:sanctum',isCompanyOwner::class]);
 });
